// src/lib/types.ts
export type JobItem = {
  id: number;
  badgeLetters: string;
  title: string;
  company: string;
  date: string;
  relevanceScore: number;
  daysAgo: number;
}

export type JobItemExpanded = JobItem & {
	description: string;
	qualifications: string[];
	reviews: string[];
	duration: string;
	location: string;
	salary: string;
	coverImgURL: string;
	companyURL: string;
}

export type PageDirection = "next" | "prev";
export type SortBy = "relevant" | "recent";